+++ src/fb.h
@@ -45,7 +45,7 @@ void fb_put(char c);
 void fb_print(const char *c);
 
 void fb_putdec(int32_t i);
-void fb_puthex(unsigned long h);
+void fb_puthex(unsigned long long h);
 
 void fb_update_timer(uint32_t tick);
